아 하다가 영 시간안에 못할거같고 해서 포기함

+++ Assets/_Script/Player/PlayerHUD.cs
@@ -15,7 +15,7 @@ public class PlayerHUD : MonoBehaviour
     public Image staminaFill;
 
     int maxHP;
-    int maxStamina;
+    float maxStamina;
 
     void Start()
     {
